requirements.txt
# UI and Core Dependencies
PyQt5>=5.15.9
PyQtWebEngine>=5.15.6  # Required for web-based Live2D avatar
pyttsx3>=2.90  # Fallback TTS engine

# AI and API Dependencies
google-generativeai>=0.8.5  # Updated to latest for GenerativeModel support
requests>=2.31.0

# Audio Processing (OpenVoice Core)
torch>=2.0.0  # Deep learning framework
torchaudio>=2.0.0  # Audio tensor operations
numpy>=1.24.0  # Numerical computations
soundfile>=0.12.1  # Audio file I/O
librosa>=0.10.1  # Audio analysis and processing

# Text Processing (OpenVoice Text Pipeline)
# Optional dependencies - will gracefully degrade if not available
# pypinyin>=0.44.0  # Chinese text processing
# cn2an>=0.5.17  # Chinese number conversion  
# jieba>=0.42.1  # Chinese word segmentation
# phonemizer>=3.2.1  # Phoneme conversion (requires espeak)

# Optional Audio Enhancement
# pyloudnorm>=0.1.1  # Audio normalization
# scipy>=1.11.0  # Signal processing utilities
